feat(STOCK-51) - Naming and structure refactoring - Added logging
